Add percentage count
While a generated report shows the total number of lines for each branch as well as lines from each author, a percentage display would be more helpful for analysing contributions during team projects.

Let's calculate the percentage of lines in the commit history belonging to each author. This makes visualisation of how much each author has contributed easier and saves the effort needed to manually calculate based on the number of lines.